Switch to libtool_check_unique from bazel
We had to include this file in this repo before a version of bazel was
released that included it. It is included with bazel 5.0 at this point,
and has received fixes that haven't been ported here. We could continue
to duplicate them here, but it doesn't seem like there's a reason to
until we try to move things like this out of bazel.

https://github.com/bazelbuild/bazel/commit/48dd159808c5a874c4aabcd27d6e66610d41bb5a