Use Bazel-provided auto-configured toolchains
The auto-configured toolchains provided by Bazel itself have diverged
heavily from the version maintained in this repo. There is no need to
maintain this duplication until Starlarkification has progressed
succiciently for rules_cc to be the source of truth for them.

This is particularly relevant for Bzlmod, which currently uses the
toolchains defined in rules_cc. As a result, Bazel C++ builds will use
subtly different toolchains depending on whether `--enable_bzlmod` is
used or not. This is fixed by loading toolchain detection logic from
`@bazel_tools` in the module extension.