Add partial support for external_include_paths
external_include_paths is an interesting feature + variable, that when
enabled, puts include paths into a separate external_include_paths
variable (separate as in not included in the default includes features).
It's also notably not in the legacy features java code. With this change
it is never activated unless users manually create the feature and add
it to their know features.