Deduplicate capability features
Under the covers, features are used to communicate tool capabilities to
the underlying code.  If multiple tools provide the same capabilities,
this ends up generating duplicate features and failing.

C++ and C tools is a really easy example, and they both have
"supports_dynamic_linker" as a capability very reasonably.

Signed-off-by: Austin Schuh <austin.linux@gmail.com>