Enable new compile_flags by default
This wasn't enabled when it was originally merged since we didn't have
the backfill feature at the time. This set of flags is susceptible to
ordering surprises so it should always be last in this feature list.
This enables users to override flags that come from the toolchain in
`copts = [...]` or with `--copt`.