rules_cc: FreeBSD 10.x and later use `libc++`, not `libstdc++`.
<https://wiki.freebsd.org/NewC++Stack> notes that FreeBSD 10.x (Jan 2014) and
onwards switched from the old GNU `libstdc++` to LLVM `libc++`, in line with the
rest of the base system switching toolchains from GNU to LLVM.

In fact, this was previously fixed by bazelbuild/bazel#14860, but apparently
that was after the great partitioning of `rules_cc`, and _this_ half of the
change got lost when we started using those build files instead.  The result is
that the `cc_binary` link step of a C++ binary will fail for the lack of a C++
standard library implementation.

This might fix the observable effects of bazelbuild/rules_cc#161 --- though that
issue correctly notes that a C compiler should not be a C++ linker, which is how
we have ended up even dealing with this problem.

PiperOrigin-RevId: 764561487
Change-Id: I535b432c90d93339fd7a8b6b70cf2a1a0d854bc8