Make windows toolchain configurable with BAZEL_*OPTS
This brings the same functionality that exists in Linux to Windows.
When cross compiling with transitions, --copts and friends apply to
both the host OS, and the target of cross compilation.  This
necessitates moving all flags to the toolchain.  There is no way to do
this today for the windows toolchains without writing a toolchain from
scratch, which isn't ideal for a couple of small flag changes.

While we are here, make `/D_WIN32_WINNT=0x0601` configurable too in the
same way, with `BAZEL_WIN32_WINNT`.