Add an allowlist_absolute_include_directories parameter to cc_args
Apple toolchain declaration in bazel are inherently non-hermetic from a
bazel perspective, as they rely on the system's Xcode installation. As a
result, apple_support currently make use of
cxx_builtin_include_directories to add known system directories, as
seen in the toolchain declaration here:
https://github.com/bazelbuild/apple_support/blob/master/crosstool/osx_cc_configure.bzl#L136

This currently cannot be done in rules based toolchain, as the
allowlist_include_directories argument in cc_args() requires a
DirectoryInfo provider.

In this change, we're adding a new parameter to cc_args() that can be
used to add arbitrary directories to cxx_builtin_include_directories.
The content will be propagated in all intermediate providers all the
way to the top toolchain definition.

Fixes #242