Add requires_not_none support for cc_args.env
Since this commit
https://github.com/bazelbuild/rules_cc/commit/5a8cab742c3868ae5fc813745c8677f676444cf1
bazel has supported expand_if_available for env_sets. This is useful
when it's easier in the tool being run to accept optional arguments via
environment variables vs potentially double parsing and filtering
arguments.

For example in the case of parse_headers actions you either have to
parse the command line (including optional response files) in order to
find the value passed to `-o`, or you can organize it so you can read
it from the environment instead with this.