Remove allowlist_include_directories from data
When very recent versions of bazel, improvements in the merkle tree
recommend that instead of passing a huge list of files to the toolchain,
you instead pass the root directory of those files once, and then bazel
symlinks the entire directory as opposed to every file one by one. In
order to do this you must pass a filegroup pointing at only the
directory. In this case, if you then also pass all the files, you get
conflicts in bazel trying to link the files, and the directory. If you
didn't get those conflicts you would still lose the perf benefits.

This change makes allowlist_include_directories only affect the
cxx_builtin_include_directories field. If you need the directory also in
data, you must now also add it to `data`. If you want to continue using
the entire list of files, that still works fine.