Prefix cxx_builtin_include_directories with %workspace%
When passing a directory as a sysroot, the files in the sysroot must be
added here. Previously just the raw relative path such as
`external/sysroot/` was set, which bazel still errored on with
undeclared inclusions. By prefixing this with `%workspace%` bazel is
happy.

Since we're always passing some bazel target here, this seems safe.